% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{browse_questionnaire}
\alias{browse_questionnaire}
\title{Browse a CIS study questionnaire}
\usage{
browse_questionnaire(study_code)
}
\arguments{
\item{study_code}{A string with the study code.}
}
\description{
Opens a pdf url with the questionnaire of a specific CIS study.
}
