% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{download_file}
\alias{download_file}
\title{Download and read a CIS study from a given URL}
\usage{
download_file(url, destfile = tempfile(fileext = ".zip"))
}
\arguments{
\item{url}{A string with the URL of the CIS study page.}

\item{destfile}{A string with the path where the ZIP file will be saved.
Defaults to a temporary file.}
}
\description{
Download and read a CIS study from a given URL
}
\keyword{internal}
