% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{downloadable_files}
\alias{downloadable_files}
\title{Retrieve the downloadable of a CIS study.}
\usage{
downloadable_files(study_code)
}
\arguments{
\item{study_code}{A string with the study code.}
}
\value{
A data.frame.
}
\description{
Retrieves a list of files that can be downloaded asociated with a specific CIS study.
This includes pdf and html files (such as questionnaires, frecuency reports or methodogical notes)
and zip files with the individual level data (usually in SPSS sav format, but not always).
}
