% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{flatten_series}
\alias{flatten_series}
\title{Flattens the timeseries response from the CIS API.}
\usage{
flatten_series(res_content)
}
\arguments{
\item{res_content}{A list with the parsed response.}
}
\value{
A data.frame with the parsed response.
}
\description{
Flattens the timeseries response from the CIS API.
}
\keyword{Internal}
