% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-categories.R
\name{get_question_categories}
\alias{get_question_categories}
\title{Retrieve the categories of CIS questions.}
\usage{
get_question_categories()
}
\value{
A data.frame.
}
\description{
Retrieves the valid categories of CIS questions.
The values of the `codigo` column can be used in the `descriptor_code`
parameter of the `search_questions()` function.
}
