% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-categories.R
\name{get_series_category}
\alias{get_series_category}
\title{Retrieve the categories of CIS timeseries.}
\usage{
get_series_category()
}
\value{
A data.frame.
}
\description{
Retrieves the valid categories of CIS timeseries.
The values of the `dmindex` column can be used in the `subject_code`
parameter of the `search_series()` function. The values of the `dmvariable`
columns can be used in the `series_code`, `since_series_code`, and `until_series_code`
parameters of the `search_series()`.
}
