% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-categories.R
\name{get_study_categories}
\alias{get_study_categories}
\title{Retrieve the categories of CIS studies.}
\usage{
get_study_categories()
}
\value{
A data.frame.
}
\description{
Retrieves the valid categories of CIS studies.
The values of the \code{codigo} column can be used in the \code{category_code}
parameter of the \code{search_studies()} function.
}
