% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{get_study_url}
\alias{get_study_url}
\title{Get the URL of a CIS study}
\usage{
get_study_url(study_code)
}
\arguments{
\item{study_code}{A string with the study ID.}
}
\value{
A string with the URL of the study, or NULL if not found.
}
\description{
Retrieves the URL of a specific CIS study using its study ID.
}
\keyword{internal}
