% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opencis-package.R
\docType{package}
\name{opencis-package}
\alias{opencis}
\alias{opencis-package}
\title{opencis: Import Data from Spanish Sociological Research Center (CIS)}
\description{
Search and import data directly to R from the [Spanish Sociological Research Center (CIS)](https://www.cis.es/inicio). The CIS is a public institution that conducts electoral and sociological research studies on the Spanish society. The CIS has a large database of surveys that can be accessed through its website. The package includes functions to search for surveys, survey questions and timeseries, and import the data directly to R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://hmeleiro.github.io/opencis/}
}

}
\keyword{internal}
