% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{parse_question}
\alias{parse_question}
\title{Parse CIS question search results}
\usage{
parse_question(resp)
}
\arguments{
\item{resp}{The HTTP response object from the CIS search.}
}
\value{
A tibble with the parsed data series information.
}
\description{
Parse CIS question search results
}
\keyword{internal}
