% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{parse_response}
\alias{parse_response}
\title{General parser for responses from the CIS API.}
\usage{
parse_response(res)
}
\arguments{
\item{res}{A response object from a httr request.}
}
\value{
A data.frame with the parsed response.
}
\description{
General parser for responses from the CIS API.
}
\keyword{Internal}
