% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{parse_response_series}
\alias{parse_response_series}
\title{Specific parser of response from the /cis/serie/ endpoint of the CIS API.}
\usage{
parse_response_series(res)
}
\arguments{
\item{res}{A response object from a httr request.}
}
\value{
A data.frame with the parsed response.
}
\description{
Specific parser of response from the /cis/serie/ endpoint of the CIS API.
}
\keyword{Internal}
