% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{parse_serie}
\alias{parse_serie}
\title{Parse CIS data series search results}
\usage{
parse_serie(resp)
}
\arguments{
\item{resp}{The HTTP response object from the CIS search.}
}
\value{
A tibble with the parsed data series information.
}
\description{
Parse CIS data series search results
}
\keyword{internal}
