% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{parse_study}
\alias{parse_study}
\title{Parse CIS study search results}
\usage{
parse_study(resp)
}
\arguments{
\item{resp}{The HTTP response object from the CIS search.}
}
\value{
A tibble with the parsed studies information.
}
\description{
Parse CIS study search results
}
\keyword{internal}
