% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_cis}
\alias{read_cis}
\title{Import a CIS study}
\usage{
read_cis(study_code)
}
\arguments{
\item{study_code}{A string with the study code.}
}
\value{
A data.frame with the study data.
}
\description{
Download and import the data of a CIS study.
}
\examples{
if (interactive()) {
# If you know the study code you can just read it into R
df <- read_cis("3328")
print(df)

# If you dont know the study code, you can search for a study using search_cis() function:
studies <- search_cis(q = "gastronomia")
print(studies)

df <- read_cis(studies$study[1])
print(df)
}
}
