% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_sav_from_zip}
\alias{read_sav_from_zip}
\title{Read a SAV file from a ZIP archive}
\usage{
read_sav_from_zip(zip_path)
}
\arguments{
\item{zip_path}{A string with the path to the ZIP file.}
}
\value{
A data.frame with the data read from the .sav file.
}
\description{
Extracts and reads the SPSS (.sav) data file contained in a ZIP archive.
}
\keyword{internal}
