% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_series}
\alias{read_series}
\title{Import a CIS timeseries}
\usage{
read_series(series_code)
}
\arguments{
\item{series_code}{A string with the timeseries code.}
}
\value{
A data.frame.
}
\description{
Download and import the data of a CIS timeseries.
}
\examples{
\dontshow{httr::set_config(httr::config(ssl_verifypeer = 0L))}
# Use the search_series() function to find a timeserie to import
series <- search_series("presidente preferido")
series
# Use the value in the id column of the search_series() output
df <- read_series(series_code = series$id[1])
dplyr::glimpse(df)
}
