% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_series_microdata}
\alias{read_series_microdata}
\title{Import the individual-level data for CIS timeseries}
\usage{
read_series_microdata(series_code, since_date = NULL, until_date = NULL)
}
\arguments{
\item{series_code}{A string (or a character vector) with the timeseries codes.}

\item{since_date}{A string with the start date of the search in '\%d-\%m-\%Y' format.}

\item{until_date}{A string with the end date of the search in '\%d-\%m-\%Y' format.}
}
\value{
A data.frame.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}} Download and import the individual-level data for a CIS timeseries.
}
