% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_series}
\alias{search_series}
\title{Search for CIS timeseries}
\usage{
search_series(
  search_terms = NULL,
  subject_code = NULL,
  since_date = NULL,
  until_date = NULL,
  series_code = NULL,
  since_series_code = NULL,
  until_series_code = NULL
)
}
\arguments{
\item{search_terms}{A string with the search terms.}

\item{subject_code}{A string (or a character vector if more than one) with the subject code. The subject code corresponds to the values of the `dmindex` column in the data.frame output of the `get_series_category()` function.}

\item{since_date}{A string with the start date of the search in \%d-\%m-\%Y format.}

\item{until_date}{A string with the end date of the search in \%d-\%m-\%Y format.}

\item{series_code}{A string with the series code wanted. The series code corresponds to the values of the `dmvariable` column in the data.frame output of the `get_series_category()` function.}

\item{since_series_code}{A string with the first series code wanted. The series code corresponds to the values of the `dmvariable` column in the data.frame output of the `get_series_category()` function.}

\item{until_series_code}{A string with the last series code wanted. The series code corresponds to the values of the `dmvariable` column in the data.frame output of the `get_series_category()` function.}
}
\value{
A data.frame.
}
\description{
Searches for CIS timeseries using the CIS search engine.
}
