% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{download_study}
\alias{download_study}
\title{Download and read a CIS study from a given URL}
\usage{
download_study(url)
}
\arguments{
\item{url}{A string with the URL of the CIS study page.}
}
\description{
Download and read a CIS study from a given URL
}
\keyword{internal}
