% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-categories.R
\name{get_series_category}
\alias{get_series_category}
\title{Retrieve the categories of CIS timeseries.}
\usage{
get_series_category()
}
\value{
A data.frame.
}
\description{
Retrieves the valid categories of CIS timeseries.
The values of the \code{dmindex} column can be used in the \code{subject_code}
parameter of the \code{search_series()} function. The values of the \code{dmvariable}
columns can be used in the \code{series_code}, \code{since_series_code}, and \code{until_series_code}
parameters of the \code{search_series()}.
}
