% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-categories.R
\name{get_study_categories}
\alias{get_study_categories}
\title{Retrieve the categories of CIS studies.}
\usage{
get_study_categories()
}
\value{
A data.frame.
}
\description{
Retrieves the valid categories of CIS studies.
The values of the `codigo` column can be used in the `category_code`
parameter of the `search_studies()` function.
}
