% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-categories.R
\name{get_study_colections}
\alias{get_study_colections}
\title{Retrieve the collection categories of CIS studies.}
\usage{
get_study_colections()
}
\value{
A data.frame.
}
\description{
Retrieves the valid collection categories of CIS studies.
The values of the `codigo` column can be used in the `collection_code`
parameter of the `search_studies()` of `search_questions()` functions.
}
