% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_cis}
\alias{read_cis}
\title{Import a CIS study}
\usage{
read_cis(study_code)
}
\arguments{
\item{study_code}{A string with the study code.}
}
\value{
A data.frame.
}
\description{
Download and import the data of a CIS study.
}
