% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.R
\name{read_series}
\alias{read_series}
\title{Import a CIS timeseries}
\usage{
read_series(series_code)
}
\arguments{
\item{series_code}{A string with the timeseries code.}
}
\value{
A data.frame.
}
\description{
Download and import the data of a CIS timeseries.
}
