% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_questions}
\alias{search_questions}
\title{Search for CIS questions.}
\usage{
search_questions(
  search_terms = NULL,
  descriptor_code = NULL,
  collection_code = NULL,
  since_date = NULL,
  until_date = NULL,
  study_code = NULL,
  since_study_code = NULL,
  until_study_code = NULL,
  scope_code = NULL
)
}
\arguments{
\item{search_terms}{A string with the search terms.}

\item{descriptor_code}{A string (or a character vector if more than one) with the descriptor code. The descriptor code corresponds to the values of the \code{codigo} column in the data.frame output of the \code{get_question_categories()} function.}

\item{collection_code}{A string (or a character vector if more than one) with the collection code. The collection code corresponds to the values of the \code{codigo} column in the data.frame output of the \code{get_study_colections()} function.}

\item{since_date}{A string with the start date of the search in '\%d-\%m-\%Y' format.}

\item{until_date}{A string with the end date of the search in '\%d-\%m-\%Y' format.}

\item{study_code}{A string with the study code wanted.}

\item{since_study_code}{A string with the first study code wanted.}

\item{until_study_code}{A string with the last study code wanted.}

\item{scope_code}{Scope level code of the survey sample ("001" for national level, "002" for autonomical comunity level, etc...)}
}
\value{
A data.frame.
}
\description{
Searches for CIS questions using the CIS search engine.
}
\examples{
\dontshow{if(Sys.info()["sysname"] == "Linux") httr::set_config(httr::config(ssl_verifypeer = 0L))}
# Search by search terms
questions <- search_questions(search_terms = "feminismo")
print(questions)

# Narrow the search by dates
questions <- search_questions(search_terms = "feminismo", since_date = "01-01-2010")
print(questions)
}
