% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_studies}
\alias{search_studies}
\title{Search for CIS studies.}
\usage{
search_studies(
  search_terms = NULL,
  category_code = NULL,
  collection_code = NULL,
  since_date = NULL,
  until_date = NULL,
  study_code = NULL,
  since_study_code = NULL,
  until_study_code = NULL,
  scope_code = NULL
)
}
\arguments{
\item{search_terms}{A string with the search terms.}

\item{category_code}{A string (or a character vector if more than one) with the category code. The category code corresponds to the values of the \code{codigo} column in the data.frame output of the \code{get_study_categories()} function.}

\item{collection_code}{A string (or a character vector if more than one) with the collection code. The collection code corresponds to the values of the \code{codigo} column in the data.frame output of the \code{get_study_colections()} function.}

\item{since_date}{A string with the start date of the search in '\%d-\%m-\%Y' format.}

\item{until_date}{A string with the end date of the search in '\%d-\%m-\%Y' format.}

\item{study_code}{A string with the study code wanted.}

\item{since_study_code}{A string with the first study code wanted.}

\item{until_study_code}{A string with the last study code wanted.}

\item{scope_code}{Scope level code of the survey sample ("001" for national level, "002" for autonomical comunity level, etc...)}
}
\value{
A data.frame.
}
\description{
Searches for CIS studies using the CIS search engine.
}
\examples{
\dontshow{if(Sys.info()["sysname"] == "Linux") httr::set_config(httr::config(ssl_verifypeer = 0L))}
# Search by search terms
studies <- search_studies(search_terms = "postelectoral")
print(studies)

# Narrow the search by dates
studies <- search_studies(search_terms = "postelectoral",
                          since_date = "01-01-2011",
                          until_date = "01-01-2020")
print(studies)

# Use the scope_code parameter to search only national studies
studies <- search_studies(search_terms = "postelectoral",
                          since_date = "01-01-2011",
                          until_date = "01-01-2020",
                          scope_code = "001")
print(studies)
}
